Reorganiza frontend y backend
