+++ frontend/css/components/search.css
